# frontend/specs/001-faq-system-upgrade/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: FAQ System API
  description: |
    DIY Ski FAQ System Backend API - 提供搜尋、Intent 偵測、Slot 擷取、分析和 CRM 整合功能

    **核心功能**:
    - FAQ 搜尋與推薦
    - Intent 偵測（Rule Engine + LLM 增強）
    - Slot 擷取（雪場、日期、人數等）
    - 搜尋分析與統計
    - CRM 整合介面

    **認證**: 目前版本不需要認證（內部使用），未來可加入 API Key
  version: 1.0.0
  contact:
    name: DIY Ski CRM Team
    url: https://diy.ski
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.diy.ski/api/v1
    description: Production server (未來部署)
...
paths:
  /faq/search:
    post:
      tags:
        - FAQ
      summary: 搜尋 FAQ
      description: |
        使用 Fuse.js 進行 FAQ 搜尋，回傳最相關的結果。
        支援多語言查詢（中文、英文、泰文）。

        **效能目標**: <100ms (P95)
      operationId: searchFAQ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  minLength: 1
                  maxLength: 100
                  description: 使用者查詢文字
                  example: "去野澤要準備什麼裝備"
                language:
                  type: string
                  enum: [zh, en, th]
                  default: zh
                  description: 查詢語言
                limit:
                  type: integer
                  minimum: 1
                  maximum: 20
                  default: 5
                  description: 最多回傳幾筆結果
                threshold:
                  type: number
                  minimum: 0
                  maximum: 1
                  default: 0.6
                  description: Fuse.js 分數閾值（0-1，越小越嚴格）
                include_related:
                  type: boolean
                  default: false
                  description: 是否包含相關 FAQ
      responses:
        '200':
          description: 搜尋成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                  - meta
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: '#/components/schemas/FAQSearchResult'
                      total:
                        type: integer
                        description: 符合條件的總數
                        example: 5
                  meta:
                    $ref: '#/components/schemas/ResponseMeta'
              example:
                success: true
                data:
                  results:
                    - faq_id: "faq.gear.001"
                      canonical_question: "去滑雪需要準備什麼裝備？"
                      answer_preview: "基本裝備包含雪衣、雪褲、手套..."
                      score: 0.12
                      confidence: 88
                      matched_fields: ["canonical_question", "keywords"]
                  total: 5
                meta:
                  timestamp: "2025-01-15T14:30:00.123Z"
                  query_id: "q_20250115_143000"
                  response_time_ms: 45
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /intent/detect:
    post:
      tags:
        - Intent
      summary: 偵測查詢意圖
      description: |
        使用 Rule Engine + LLM 混合方式偵測使用者意圖。

        **決策邏輯**:
        - Rule Engine 信心度 ≥ 75%: 直接使用 Rule Engine 結果
        - Rule Engine 信心度 < 75%: 呼叫 LLM API 增強

        **效能目標**:
        - Rule Engine: <50ms
        - LLM 增強: <2000ms
      operationId: detectIntent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  minLength: 1
                  maxLength: 100
                  example: "2月15日去野澤，2大1小"
                use_llm_fallback:
                  type: boolean
                  default: true
                  description: 當 Rule Engine 信心度低時是否使用 LLM
                include_slots:
                  type: boolean
                  default: true
                  description: 是否同時執行 Slot 擷取
      responses:
        '200':
          description: 偵測成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/IntentDetectionResult'
                  meta:
                    $ref: '#/components/schemas/ResponseMeta'
              example:
                success: true
                data:
                  query_text: "2月15日去野澤，2大1小"
                  detected_intent: "ITINERARY"
                  confidence: 85
                  method: "rule_engine"
                  rule_engine:
                    score: 85
                    breakdown:
                      intent_match: 35
                      keyword_match: 25
                      slot_match: 18
                      historical_ctr: 7
                    matched_keywords: ["野澤", "2大1小", "2月"]
                  slots:
                    resort:
                      code: "NOZAWA"
                      name_cn: "野澤溫泉"
                      matched_text: "野澤"
                    date:
                      date: "2025-02-15"
                      matched_text: "2月15日"
                    people:
                      adults: 2
                      children: 1
                      matched_text: "2大1小"
                meta:
                  timestamp: "2025-01-15T14:30:00.123Z"
                  query_id: "q_20250115_143000"
                  response_time_ms: 38
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
    FAQSearchResult:
      type: object
      required:
        - faq_id
        - canonical_question
        - score
        - confidence
      properties:
        faq_id:
          type: string
          example: "faq.gear.001"
        canonical_question:
          type: string
          description: 標準問題（依語言參數回傳對應語言）
          example: "去滑雪需要準備什麼裝備？"
        answer_preview:
          type: string
          description: 答案預覽（前 100 字元）
          example: "基本裝備包含雪衣、雪褲、手套..."
        score:
          type: number
          description: Fuse.js 評分（0-1，越小越好）
          example: 0.12
        confidence:
          type: number
          description: 信心度（0-100）
          example: 88
        matched_fields:
          type: array
          items:
            type: string
          description: 匹配的欄位
          example: ["canonical_question", "keywords"]
        related_faqs:
          type: array
          items:
            type: object
            properties:
              faq_id:
                type: string
              canonical_question:
                type: string
...
    IntentDetectionResult:
      type: object
      required:
        - query_text
        - detected_intent
        - confidence
        - method
      properties:
        query_text:
          type: string
        detected_intent:
          type: string
          enum: [ITINERARY, BOOKING, COURSE, INSTRUCTOR, KIDS_SAFETY, PAYMENT, GEAR, SERVICE, GENERAL]
        confidence:
          type: number
          description: 0-100
        method:
          type: string
          enum: [rule_engine, llm, hybrid]
        rule_engine:
          type: object
          properties:
            score:
              type: number
            breakdown:
              type: object
              properties:
                intent_match:
                  type: number
                keyword_match:
                  type: number
                slot_match:
                  type: number
                historical_ctr:
                  type: number
            matched_keywords:
              type: array
              items:
                type: string
        llm_enhancement:
          type: object
          properties:
            intent:
              type: string
            confidence:
              type: number
            reasoning:
              type: string
            model:
              type: string
            token_usage:
              type: object
              properties:
                input:
                  type: integer
                output:
                  type: integer
            latency_ms:
              type: number
        slots:
          $ref: '#/components/schemas/SlotExtractionResult'
    SlotExtractionResult:
      type: object
      properties:
        query_text:
          type: string
        slots:
          type: object
          properties:
            resort:
              type: object
              properties:
                code:
                  type: string
                  example: "NOZAWA"
                name_cn:
                  type: string
                name_en:
                  type: string
                name_jp:
                  type: string
                matched_text:
                  type: string
            date:
              type: object
              properties:
                date:
                  type: string
                  format: date
                date_range:
                  type: object
                  properties:
                    start:
                      type: string
                      format: date
                    end:
                      type: string
                      format: date
                matched_text:
                  type: string
            people:
              type: object
              properties:
                adults:
                  type: integer
                children:
                  type: integer
                matched_text:
                  type: string
            board_type:
              type: string
              enum: [SKI, SNOWBOARD]
            child_age:
              type: array
              items:
                type: integer
        confidence:
          type: object
          description: 每個 Slot 的信心度（0-100）
          additionalProperties:
            type: number
        extraction_method:
          type: object
          description: 每個 Slot 的擷取方法
          additionalProperties:
            type: string
...
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2025-01-15T14:30:00.123Z"
        query_id:
          type: string
          example: "q_20250115_143000"
        response_time_ms:
          type: integer
          example: 45
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - meta
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: "INVALID_QUERY"
            message:
              type: string
              example: "查詢文字不得為空"
            details:
              type: object
              description: 詳細錯誤資訊（僅開發模式）
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            query_id:
              type: string
  responses:
    BadRequest:
      description: 請求參數錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "INVALID_QUERY"
              message: "查詢文字不得為空"
            meta:
              timestamp: "2025-01-15T14:30:00.123Z"
...
    InternalError:
      description: 伺服器內部錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "INTERNAL_ERROR"
              message: "伺服器發生錯誤，請稍後再試"
            meta:
              timestamp: "2025-01-15T14:30:00.123Z"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key（未來版本啟用）

# 目前版本不需要認證
security: []